refactor(tests): move history tests closer to implementation
